jadx-0.7.2 - B2 v4.8.8 (1255)

+++ sources/org/telegram/messenger/voip/VoIPController.java
@@ -10,6 +10,7 @@ import java.util.Calendar;
 import java.util.Iterator;
 import java.util.Locale;
 import org.telegram.messenger.ApplicationLoader;
+import org.telegram.messenger.BuildConfig;
 import org.telegram.messenger.MessagesController;
 import org.telegram.tgnet.TLRPC.TL_phoneConnection;
 import org.telegram.ui.Components.voip.VoIPHelper;
@@ -233,6 +234,8 @@ public class VoIPController {
     }
 
     public void setConfig(double recvTimeout, double initTimeout, int dataSavingOption, long callID) {
+        String logFilePath;
+        String logFilePath2;
         ensureNativeInstance();
         boolean sysAecAvailable = false;
         boolean sysNsAvailable = false;
@@ -247,7 +250,17 @@ public class VoIPController {
         long j = this.nativeInst;
         boolean z = (sysAecAvailable && VoIPServerConfig.getBoolean("use_system_aec", true)) ? false : true;
         boolean z2 = (sysNsAvailable && VoIPServerConfig.getBoolean("use_system_ns", true)) ? false : true;
-        nativeSetConfig(j, recvTimeout, initTimeout, dataSavingOption, z, z2, true, getLogFilePath(callID), null);
+        if (BuildConfig.DEBUG) {
+            logFilePath = getLogFilePath("voip");
+        } else {
+            logFilePath = getLogFilePath(callID);
+        }
+        if (BuildConfig.DEBUG && dump) {
+            logFilePath2 = getLogFilePath("voipStats");
+        } else {
+            logFilePath2 = null;
+        }
+        nativeSetConfig(j, recvTimeout, initTimeout, dataSavingOption, z, z2, true, logFilePath, logFilePath2);
     }
 
     public void debugCtl(int request, int param) {
@@ -284,20 +297,22 @@ public class VoIPController {
 
     private String getLogFilePath(long callID) {
         File dir = VoIPHelper.getLogsDir();
-        File[] _logs = dir.listFiles();
-        ArrayList<File> logs = new ArrayList();
-        logs.addAll(Arrays.asList(_logs));
-        while (logs.size() > 20) {
-            File oldest = (File) logs.get(0);
-            Iterator it = logs.iterator();
-            while (it.hasNext()) {
-                File file = (File) it.next();
-                if (file.getName().endsWith(".log") && file.lastModified() < oldest.lastModified()) {
-                    oldest = file;
+        if (!BuildConfig.DEBUG) {
+            File[] _logs = dir.listFiles();
+            ArrayList<File> logs = new ArrayList();
+            logs.addAll(Arrays.asList(_logs));
+            while (logs.size() > 20) {
+                File oldest = (File) logs.get(0);
+                Iterator it = logs.iterator();
+                while (it.hasNext()) {
+                    File file = (File) it.next();
+                    if (file.getName().endsWith(".log") && file.lastModified() < oldest.lastModified()) {
+                        oldest = file;
+                    }
                 }
+                oldest.delete();
+                logs.remove(oldest);
             }
-            oldest.delete();
-            logs.remove(oldest);
         }
         return new File(dir, callID + ".log").getAbsolutePath();
     }
